disambiguated things a ton by splitting the parser into ValueParser
and AssemblyParser. Basically, there are only ever certain places values
can be, and it's the same with opcodes and label definitions. So it
makes sense to look up values, opcodes, and instructions all
differently.

This simplifies a huge amount of things, too. I'll clean up more once I
finish adapting the tests.