implemented a conceptually simple parsing framework, nicer than the last
its based around two idea: regular expression matching and Defer.

Any parse function can raise Defer to tell the parser that it doesn't
know how to parse the thing it was given. This is wrapped up in a nice
api that lets us use a decorator to assign each parse function a regular
expression. Then it wraps each parse function in a closure that raises a
Defer if its regular expression doesn't match. Neat, huh?