simplified everything quite a bit -- pass around a state.
before we were passing around a thousand little locals. No longer! Now
we have convenient wrappers for keeping everything transactional, and
all is well. The state gets passed around; each value keeps a reference
to it and can modify it on its turn. Finally, if the cpu wants to, those
changes are applied to the cpu itself. wooooooh.

It's the nicest thing ever when all your code collapses once you
discover a simple abstraction.